Switching SearchPanel items retrieval to dynamic instead of static on every render
Rendering time was ~115ms even when only displaying 10 items max -- it was down to only ~15 when using the search with a much smaller datastore.
The lag is probably because SearchPanel re-renders every time a letter is typed, even when no item is rendered (below threshold), and so it triggers a call to getQueriesData every single time.