chore: remove unnecessary TODOs
Extension class should only give access to data if provided with the whole item, for the time being. This is to limit access to Zotero metadata outside of the extension's functionalities itself.

The one exception is bibliographic entries, which is used by the LaTeX exporter.

Can be revisited in the future, except potentially for methods that can return unprocessed metadata